requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.2.1
django-allauth==0.52.0
django-bootstrap-form==3.4
django-colorfield==0.8.0
django-contrib-comments==2.2.0
django-crispy-forms==2.0
django-logentry-admin==1.1.0
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-otp==0.2.7
django-rest-auth==0.9.5
django-session-security==2.6.7
django-threadedcomments==2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
frozenlist==1.3.3
idna==3.4
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.17.3
multidict==6.0.4
mypy-extensions==1.0.0
mysql-client==0.0.1
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.3
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
twilio==8.1.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.1
yarl==1.9.2
